Move cmk.base.sources._checkers -> cmk.base.source
Package with a single file -> module
